requirements.txt
# =============================================================================
# DEPENDENCIAS PRINCIPALES
# =============================================================================

# Streamlit para la interfaz web
streamlit>=1.28.0

# Google Cloud Services
google-cloud-storage>=2.10.0
google-cloud-functions>=1.13.0
google-cloud-secret-manager>=2.16.0
functions-framework>=3.4.0

# OpenAI API
openai>=1.3.0

# Procesamiento de PDF
marker-pdf>=0.2.0

# Vectorización y búsqueda
faiss-cpu>=1.7.4
# faiss-gpu>=1.7.4  # Descomenta si usas GPU

# Análisis de datos
pandas>=2.0.0
numpy>=1.24.0

# Utilidades
python-dotenv>=1.0.0
pydantic>=2.0.0
tqdm>=4.65.0
pathlib>=1.0.1

# Validación de archivos y seguridad
python-magic>=0.4.27

# Logging y monitoreo
logging>=0.4.9.6

# Desarrollo y testing
pytest>=7.0.0
pytest-cov>=4.0.0

# =============================================================================
# DEPENDENCIAS OPCIONALES
# =============================================================================

# Para procesamiento de texto adicional
nltk>=3.8.0
spacy>=3.6.0

# Para mejor manejo de archivos
python-magic>=0.4.27

# Para optimización de memoria
psutil>=5.9.0

# Para validación de datos
pydantic>=2.0.0

# Para monitoreo avanzado
prometheus-client>=0.17.0

# =============================================================================
# DEPENDENCIAS DE DESARROLLO
# =============================================================================

# Formateo de código
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0

# Tipo checking
mypy>=1.5.0

# Documentación
mkdocs>=1.5.0
mkdocs-material>=9.2.0

# Testing y Pruebas
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
coverage>=7.3.0 